+++ resharper/resharper-unity/test/data/CSharp/Intentions/ContextActions/ConvertXmlDocToTooltipAttribute/ConvertSerialisedFieldWithXmlDoc01.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+public class MyBehaviour : MonoBehaviour
+{
+    /// <summary>
+    /// Copy this to tooltip
+    /// </summary>
+    public string MySerialised{caret}Field;
+}