+++ resharper/resharper-unity/test/data/CSharp/Intentions/ContextActions/ConvertXmlDocToTooltipAttribute/ConvertSerialisedFieldWithXmlDoc04.cs
@@ -1,0 +1,12 @@
+using UnityEngine;
+
+public class MyBehaviour : MonoBehaviour
+{
+    /// <summary>
+    /// Copy this to {caret}tooltip
+    /// </summary>
+    /// <remarks>
+    /// XML Docs should not be deleted
+    /// </remarks>
+    public string MySerialisedField;
+}