+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Since 2018.1, the version numbers and release cycle match Rider's versions and r
 ### Added
 
 - Content of `[Tooltip]` attribute shown when hovering over a field ([#2032](https://github.com/JetBrains/resharper-unity/issues/2032), [RIDER-64323](https://youtrack.jetbrains.com/issue/RIDER-64323), [#2163](https://github.com/JetBrains/resharper-unity/pull/2163))
+- Add context action to convert XML doc to `[Tooltip]` attribute for a serialised field ([RIDER-64613](https://youtrack.jetbrains.com/issue/RIDER-64613), [#2163](https://github.com/JetBrains/resharper-unity/pull/2163))
 
 ### Changed